feat(claw): update device and image

+++ lib/claw/device_get_device.go
@@ -54,8 +54,8 @@ func (s *Claw) GetDevice(ctx context.Context, req *clawv1.GetDeviceRequest) (*cl
 		ImageMinWidth:         uint32(deviceRow.ImageMinWidth),
 		ImageMaxHeight:        uint32(deviceRow.ImageMaxHeight),
 		ImageMaxWidth:         uint32(deviceRow.ImageMaxWidth),
-		ImageMinFilesize:      uint64(deviceRow.ImageMinFileSize),
-		ImageMaxFilesize:      uint64(deviceRow.ImageMaxFileSize),
+		ImageMinFilesize:      uint32(deviceRow.ImageMinFileSize),
+		ImageMaxFilesize:      uint32(deviceRow.ImageMaxFileSize),
 		Nsfw:                  clawv1.NSFWMode(deviceRow.NsfwMode),
 		CreatedAt:             deviceRow.CreatedAt.ToProto(),
 		UpdatedAt:             deviceRow.UpdatedAt.ToProto(),
@@ -66,4 +66,3 @@ func (s *Claw) GetDevice(ctx context.Context, req *clawv1.GetDeviceRequest) (*cl
 		Device: device,
 	}, nil
 }
-
