+++ lib/claw/device_list_devices.go
@@ -84,8 +84,8 @@ func (s *Claw) ListDevices(ctx context.Context, req *clawv1.ListDevicesRequest) 
 			ImageMinWidth:         uint32(deviceRow.ImageMinWidth),
 			ImageMaxHeight:        uint32(deviceRow.ImageMaxHeight),
 			ImageMaxWidth:         uint32(deviceRow.ImageMaxWidth),
-			ImageMinFilesize:      uint64(deviceRow.ImageMinFileSize),
-			ImageMaxFilesize:      uint64(deviceRow.ImageMaxFileSize),
+			ImageMinFilesize:      uint32(deviceRow.ImageMinFileSize),
+			ImageMaxFilesize:      uint32(deviceRow.ImageMaxFileSize),
 			Nsfw:                  clawv1.NSFWMode(deviceRow.NsfwMode),
 			CreatedAt:             deviceRow.CreatedAt.ToProto(),
 			UpdatedAt:             deviceRow.UpdatedAt.ToProto(),
@@ -126,4 +126,3 @@ func (s *Claw) ListDevices(ctx context.Context, req *clawv1.ListDevicesRequest) 
 		TotalCount:    totalCount,
 	}, nil
 }
-
