+++ lib/claw/device_subscribe_device.go
@@ -108,8 +108,8 @@ func (s *Claw) SubscribeDevice(ctx context.Context, req *clawv1.SubscribeDeviceR
 		ImageMinWidth:         uint32(deviceRow.ImageMinWidth),
 		ImageMaxHeight:        uint32(deviceRow.ImageMaxHeight),
 		ImageMaxWidth:         uint32(deviceRow.ImageMaxWidth),
-		ImageMinFilesize:      uint64(deviceRow.ImageMinFileSize),
-		ImageMaxFilesize:      uint64(deviceRow.ImageMaxFileSize),
+		ImageMinFilesize:      uint32(deviceRow.ImageMinFileSize),
+		ImageMaxFilesize:      uint32(deviceRow.ImageMaxFileSize),
 		Nsfw:                  clawv1.NSFWMode(deviceRow.NsfwMode),
 		CreatedAt:             deviceRow.CreatedAt.ToProto(),
 		UpdatedAt:             deviceRow.UpdatedAt.ToProto(),
@@ -120,3 +120,4 @@ func (s *Claw) SubscribeDevice(ctx context.Context, req *clawv1.SubscribeDeviceR
 		Device: device,
 	}, nil
 }
+
